pyproject.toml
[project]
name = "samantha"
version = "0.1.0"
description = "POC for evaluating LLM models on laboratory workflow routing for breast cancer specimens"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27,<1.0",
    "pyyaml>=6.0.3,<7.0",
    "chromadb>=0.5,<1.0",
    "posthog>=2.4,<4",
    "sentence-transformers>=3.0,<4.0",
    "rich>=13.0,<15.0",
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.34,<1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
    "mypy>=1.14",
    "pyyaml>=6.0.3",
    "types-pyyaml>=6.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-m 'not integration'"
markers = [
    "integration: tests requiring a running ollama instance or API key",
]
